ParseDimension method moved  to dimension utils
